Create/delete a place holder doc before/after removing (owner, repo) docs
Since daily-sync DAGs read uniq (owner, repo) pairs on the run to generate
the task list. When deleting all the docs in a (owner, repo) might make
the daily-sync DAGs lose some (owner, repo) pairs. Then airflow tries to
keep the task list according to the uniq(owner, repo) pairs, so some
running task might be terminated, which is not expected.

The solution is simple and nasty, we need a better solution.