Add the dag to sync opensearch issuetimeline data by uniq owners
In Opensearch, nested aggregations on (owner, repo) takes too long to
execute which will make the DAG non-schedulable. By reducing the
aggregation layers(only aggregate owners in DAG) takes less resource to
constructe the group and make the second layer of aggregation(on repos)
in the task run. So the DAG is much less expensive.